Renames js-new to js
