+++ src/index.js
@@ -4,6 +4,6 @@
 import './scss/tailwind.scss';
 import './scss/index.scss';
 
-import './js-new/index';
+import './js/index';
 
 /* eslint-enable no-undef, no-unused-vars */
