+++ src/js/global/window-ready.js
